+++ CHANGELOG.md
@@ -1,3 +1,44 @@
+## 3.38.0
+
+### 34 new icons:
+
+- `filled/adjustments-horizontal`
+- `filled/check`
+- `filled/chevron-down`
+- `filled/chevron-right`
+- `filled/click`
+- `filled/copy`
+- `filled/database`
+- `filled/device-floppy`
+- `filled/device-gamepad-2`
+- `filled/dots-vertical`
+- `filled/dots`
+- `filled/download`
+- `filled/edit`
+- `filled/external-link`
+- `filled/headset`
+- `filled/help`
+- `filled/home-2`
+- `filled/id`
+- `filled/link`
+- `filled/list-check`
+- `filled/list-details`
+- `filled/list`
+- `filled/menu-2`
+- `filled/messages`
+- `filled/pencil`
+- `filled/plus`
+- `filled/scale`
+- `filled/school`
+- `filled/search`
+- `filled/send`
+- `filled/ticket`
+- `filled/tools-kitchen-2`
+- `filled/world`
+- `filled/x`
+
+Fixed icons: `filled/calendar-event`, `filled/trash`
+
 ## 3.37.1+1
 
 ### Fixed